Feat. Export en XML des référentiels

+++ src/Classes/Import/ReferentielCompetenceImport.php
@@ -153,7 +153,7 @@ class ReferentielCompetenceImport
                 foreach ($sem->ressources->ressource as $ressource) {
                     $ar = new ApcRessource();
                     $ar->setSemestre($semestre);
-                    $ar->setOrdre(substr($ressource['code'], 2, 2));
+                    $ar->setOrdre((int)substr($ressource['code'], 2, 2));
                     $ar->setLibelle($ressource->titre);
                     $ar->setCodeMatiere((string)$ressource['code']);
                     $ar->setTdPpn((float)$ressource['heuresCMTD']);
@@ -181,7 +181,7 @@ class ReferentielCompetenceImport
                     $ar = new ApcSae();
                     $ar->setSemestre($semestre);
                     $ar->setLibelle($sae->titre);
-                    $ar->setOrdre(substr($sae['code'], 4, 2));
+                    $ar->setOrdre((int)substr($sae['code'], 4, 2));
                     $ar->setCodeMatiere((string)$sae['code']);
                     $ar->setTdPpn((float)$sae['heuresCMTD']);
                     $ar->setTpPpn((float)$sae['heuresTP']);
